requirements.txt
#Image Processing
Pillow==10.0.0
numpy==1.25.1
aggdraw~=1.3.16

#Website APi
fastapi==0.100.1
uvicorn==0.23.1
Jinja2==3.1.2
python-multipart==0.0.6
dill==0.3.7

#Database
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
python-dotenv==1.0.0

#Testing
mock==5.1.0